Build: Fix some issues with MSVC compiler

+++ Application/ChipTune/ChipTune.c
@@ -121,7 +121,7 @@ UefiMain (
 
   if (BeepGenProtocol != NULL) {
     Status = BeepGenProtocol->GenBeep (
-      Count,
+      (UINT32) Count,
       Signal,
       Silence
       );
